Move class to it's own file
