+++ com.unity.mobile.notifications/Runtime/iOS/QueryLastRespondedNotificationOp.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: dd15e13a959094f09927755f3e83c689
